pyproject.toml
[tool.poetry]
name = "telecode"
version = "0.2.0"
description = "Simple app for visually impaired students to recieve remote source code files."
authors = ["Ruslan Iskov <ruslan.rv.ua@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
wxPython = "^4.1.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
Markdown = "^3.3.4"
black = "^21.9b0"
requests = "^2.26.0"
gen-exe = "^0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
